Make xmalloc functions faster and more robust
Make the memory header have two size_t rather than two int.
Remove redundant locks and asserts
change malloc/memset to calloc function calls

+++ src/common/xmalloc.h
@@ -108,7 +108,7 @@ void *slurm_try_xmalloc(size_t , const char *, int , const char *);
 void slurm_xfree(void **, const char *, int, const char *);
 void *slurm_xrealloc(void **, size_t, const char *, int, const char *);
 int  slurm_try_xrealloc(void **, size_t, const char *, int, const char *);
-int  slurm_xsize(void *, const char *, int, const char *);
+size_t slurm_xsize(void *, const char *, int, const char *);
 
 #define XMALLOC_MAGIC 0x42
